Remove 01_zone 02_heart file and Create New zone2 files
