+++ run.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+python3 /root/clent_sleep_ver2/src/co2_cam.py &
+python3 /root/clent_sleep_ver2/src/heart.py &
+python3 /root/clent_sleep_ver2/src/zone2_socket.py &
